Fix a bug where the Scala code would sometimes print zeros
Because we were registering callbacks, but not actually waiting on the
futures, when both of the futures contained sleeps, the program would
exit before either on completed.

Scala being Scala, the `onComplete`s still ran, but with a `Failure(ExecutionException)` as the argument, rather than a `Success(())`.
This being minimal code, we ignored this, and printed the variable with
the initial value, since neither `Future` had set this.

Changing the code to `Await` on the `Futures`, rather than register a
callback prevents this from happening.